+++ farben/src/macros/mod.rs
@@ -8,6 +8,10 @@ pub mod eprint;
 pub mod expand;
 pub mod format;
 pub mod load;
+#[deprecated(
+    since = "0.18.0",
+    note = "use the `inline` feature, it works with the `c*` family. will be deleted at 0.19 stable"
+)]
 pub mod markdown;
 pub mod strip;