+++ testfarbenyay/src/main.rs
@@ -5,4 +5,6 @@ fn main() {
     init_styles();
     cprintln!("]]]]]] Bold red");
     // cprintln!("\\[ wouldn't work no more");
+    cprintln!("`thing` yay!");
+    cprintln!("[red]thing[/red]")
 }